RIO-6197 (Race condition - If port settings event and node stop command arrive simultaneously - omx component may fail port flush command)
Race condition may occur when port settings changed event and node stop command arrive almost simultaneously.
The node may attempt to do "port flush" command (as part of port reconfig) when the component is already in
idle state.  Port flush does not work in the component when it is in idle state (returns an error event).