Fix the stop handling when the writer thread is used.
The flush operation is not the only way to stop the composer.
The other way is by reaching a limit (size or time), in which case
an event is reported to the author engine AND the file is flushed.

This is an issue because the flush call on the writer thread now needs
to be reentrant (not called from the main thread anymore).

To move the call back on the main thread we use an extra field to hold
the event that should be fired. On the next fragment, the event is sent
and the DoStop method will be run from the main thread as a result.

Added extra code to make sure the worker thread flushes its buffer queue.
Added a bool to remember that exit was called. exitPending form Thread
gets cleared before the dtor is invoked.
In the flush method, don't wait if the flush happened.