Audio path stall in AMR RTSP streaming
With some settings of the OMX decoder, the audio path will get into
a mode where it will discard all incoming audio data until a marker
bit is set.  There is no logic to stop this behavior in the case
of AMR audio.  This patch is to set the marker bit in the case of
AMR audio to fix this behavior.