bug 2114674 [RIO-7463] RTSP amr-nb streaming can not start playback, need to seek then playback start In PVMFMediaClock the clock observer will pass its driver latency to the clock and the clock will make clock adjustments on its own inside PVMFMediaClock class rather than having each module take care of its latency.
Now on android, Audio MIO has a latency of 350 msecs, so when the clock starts, it starts with a value of "-350" msecs (PVMFMediaClock takes care of the latency of the device set by observer).

In jitterbuffer node before clock start, estimated server clock is around 4000 msecs and client clock (PVMFMediaClock) is zero (A difference of 4 secs is needed to start the playback, this is the initial amount of buffering required). After clock starts Jitterbuffer again queries for the time, here estimated server clock is again around 4000 msecs but client clock is -350 msecs. Since the queried currenttime is a uint32 variable, the value for the client clock becomes very huge (due to overflow) and is greater than estimated server clock by a very big value. Jitterbuffer node finds the condition of estimated server clock being less than client clock and goes into Underflow condition and Engine goes into Auto-Pause state. Estimated server clock keeps increasing but will not be able to catch the client clock (since it is very big) and player will remain in Underflow condition for long time and hence we see the issue.

Fix for the issue -
Use PVTimeComparsionUtils class for comparing clock values. The class takes care of -ve values of clock and gives the correct comparison.