Remove an extra call of ResetData from ThreadLogoff in AndroidSurfaceOutput. This extra call results in a crash if Reset is called during Engine's track selection failure.
Extra ResetData call from ThreadLogoff cleanups the Reset command from Command Queue in the MIO,
and this ThreadLogoff was called from the callback of command complete of Reset command of Video MIO.
When the call returns back to the Video MIO and tries to delete the Reset Command from the queue, it crashes
since the command had already been deleted because of ResetData call from ThreadLogoff.
There is no need for an extra ResetData call from ThreadLogoff, since ResetData will be called from Reset() and
Reset() will always be called before ThreadLogoff.