Fix an incorrect time profiling bug in camera MIO
The problem was that when DoStop or DoReset was called,
the first sample time stamp,iAudioFirstFrameTs, was reset right way, which
made the VF detection calculation in the writeComplete() callback incorrect.

The solution is to move the iAudioFirstFrameTs reset logic to DoStart()
to make sure that we still always has the correct initialized first
sample time stamp.