Log a potential deadlock issue in audio MIO for recording
When the number of bytes returned from audio driver is <= 0,
the existing audio MIO terminates the audio recording thread
in the MIO immediately. but when the application tries to
terminate the audio recording via stop or reset, the stop
or reset can potentially wait forever.

This is not verified, thus I enable the logging first to get
confirmation should this happens.

bug - 2484098