The file writer thread is not thread-safe.
We can run into problem where deallocate() called from two
separate thread which breaks PV's thread model where
every task is scheduled to run within a single omx thread.

I am pretty sure that this is why the free chunk available
notification is lost sometimes; however, this is certainly
not the fix to all the problems that we have seen so far.

bug - 2484098

Change-Id: I88bc494a11ffd739bc5771fa75b4643de503714f